+++ spec/cuprum/function_examples.rb
@@ -983,6 +983,16 @@ module Spec::Examples
           end # context
 
           context 'when the operation generates errors' do
+            shared_examples 'should display a warning' do
+              it 'should display a warning' do
+                allow(Cuprum).to receive(:warn)
+
+                instance.call
+
+                expect(Cuprum).to have_received(:warn).with(warning_message)
+              end # it
+            end # shared_examples
+
             let(:value)           { 'returned value'.freeze }
             let(:value_or_result) { value }
             let(:expected_errors) do
@@ -999,6 +1009,10 @@ module Spec::Examples
 
                 returned
               end # lambda
+            end # let
+            let(:warning_message) do
+              '#process returned a result, but there were already errors ' \
+              "#{expected_errors.inspect}"
             end # let
 
             it 'should return a result', :aggregate_failures do
@@ -1022,15 +1036,13 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
-                expect(result.success?).to be false
-                expect(result.failure?).to be true
+                expect(result.errors).to be_empty
+                expect(result.success?).to be true
+                expect(result.failure?).to be false
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a failing operation' \
@@ -1040,21 +1052,18 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
+                expect(result.errors).to be_empty
                 expect(result.success?).to be false
                 expect(result.failure?).to be true
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns an operation with ' \
             'errors' do
               let(:implementation_errors) { ['errors.messages.custom'] }
-              let(:expected_errors)       { super() + implementation_errors }
 
               it 'should return a result', :aggregate_failures do
                 result = instance.call
@@ -1062,7 +1071,7 @@ module Spec::Examples
                 expect(result).to be_a result_class
                 expect(result.value).to be value
 
-                [*expected_errors, *implementation_errors].each do |message|
+                implementation_errors.each do |message|
                   expect(result.errors).to include message
                 end # each
 
@@ -1070,6 +1079,8 @@ module Spec::Examples
                 expect(result.failure?).to be true
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a halted operation' do
@@ -1078,15 +1089,13 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
-                expect(result.success?).to be false
-                expect(result.failure?).to be true
+                expect(result.errors).to be_empty
+                expect(result.success?).to be true
+                expect(result.failure?).to be false
                 expect(result.halted?).to be true
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a result' do
@@ -1095,15 +1104,13 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
-                expect(result.success?).to be false
-                expect(result.failure?).to be true
+                expect(result.errors).to be_empty
+                expect(result.success?).to be true
+                expect(result.failure?).to be false
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a failing result' do
@@ -1112,21 +1119,18 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
+                expect(result.errors).to be_empty
                 expect(result.success?).to be false
                 expect(result.failure?).to be true
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a result with errors'\
             do
               let(:implementation_errors) { ['errors.messages.custom'] }
-              let(:expected_errors)       { super() + implementation_errors }
 
               it 'should return a result', :aggregate_failures do
                 result = instance.call
@@ -1134,7 +1138,7 @@ module Spec::Examples
                 expect(result).to be_a result_class
                 expect(result.value).to be value
 
-                [*expected_errors, *implementation_errors].each do |message|
+                implementation_errors.each do |message|
                   expect(result.errors).to include message
                 end # each
 
@@ -1142,6 +1146,8 @@ module Spec::Examples
                 expect(result.failure?).to be true
                 expect(result.halted?).to be false
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
 
             wrap_context 'when the implementation returns a halted result' do
@@ -1150,15 +1156,13 @@ module Spec::Examples
 
                 expect(result).to be_a result_class
                 expect(result.value).to be value
-
-                expected_errors.each do |message|
-                  expect(result.errors).to include message
-                end # each
-
-                expect(result.success?).to be false
-                expect(result.failure?).to be true
+                expect(result.errors).to be_empty
+                expect(result.success?).to be true
+                expect(result.failure?).to be false
                 expect(result.halted?).to be true
               end # it
+
+              include_examples 'should display a warning'
             end # wrap_context
           end # context
         end # shared_examples