+++ lib/cuprum/operation.rb
@@ -60,7 +60,7 @@ module Cuprum
       #   @yield If a block argument is given, it will be passed to the
       #     implementation.
       #
-      #   @raise [NotImplementedError] Unless a block was passed to the
+      #   @raise [Cuprum::NotImplementedError] Unless a block was passed to the
       #     constructor or the #process method was overriden by a Function
       #     subclass.
       #