+++ spec/cuprum/function_spec.rb
@@ -13,17 +13,6 @@ RSpec.describe Cuprum::Function do
     it { expect(described_class).to be_constructible.with(0).arguments }
   end # describe
 
-  describe '::NotImplementedError' do
-    let(:error_class) { described_class::NotImplementedError }
-
-    it { expect(described_class).to have_constant(:NotImplementedError) }
-
-    it 'should be an error class' do
-      expect(error_class).to be_a Class
-      expect(error_class).to be < StandardError
-    end # it
-  end # describe
-
   include_examples 'should implement the Function methods'
 
   include_examples 'should implement the generic Function methods'